requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
sqlalchemy
psycopg2-binary

